+++ internal/core/renderer.rs
@@ -7,7 +7,7 @@ use alloc::rc::Rc;
 use core::pin::Pin;
 
 use crate::api::PlatformError;
-use crate::graphics::SharedImageBuffer;
+use crate::graphics::{Rgba8Pixel, SharedPixelBuffer};
 use crate::item_tree::ItemTreeRef;
 use crate::items::TextWrap;
 use crate::lengths::{LogicalLength, LogicalPoint, LogicalRect, LogicalSize, ScaleFactor};
@@ -118,9 +118,9 @@ pub trait RendererSealed {
         Ok(())
     }
 
-    /// Re-implement this function to support Window::grab_window(), i.e. return
+    /// Re-implement this function to support Window::take_snapshot(), i.e. return
     /// the contents of the window in an image buffer.
-    fn screenshot(&self) -> Result<SharedImageBuffer, PlatformError> {
-        Err("WindowAdapter::grab_window is not implemented by the platform".into())
+    fn take_snapshot(&self) -> Result<SharedPixelBuffer<Rgba8Pixel>, PlatformError> {
+        Err("WindowAdapter::take_snapshot is not implemented by the platform".into())
     }
 }